refactor : 구조변경
