--- src/main/java/kr/hhplus/ecommerce/interfaces/presentation/handler/Exception.java
@@ -1,5 +1,0 @@
-package kr.hhplus.ecommerce.interfaces.presentation.handler;
-
-public class Exception {
-
-}
