+++ src/test/java/kr/hhplus/ecommerce/application/popular/PopularFacadeCacheTest.java
@@ -1,5 +1,7 @@
 package kr.hhplus.ecommerce.application.popular;
 
+import kr.hhplus.ecommerce.application.popular.dto.PopularCriteria;
+import kr.hhplus.ecommerce.application.popular.dto.PopularResult;
 import kr.hhplus.ecommerce.config.CacheType;
 import kr.hhplus.ecommerce.config.RedisCacheCleaner;
 import kr.hhplus.ecommerce.config.RedisCacheTemplate;
